ai_backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
cachetools==5.5.0
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
distlib==0.3.9
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.4
fastapi-cli==0.0.5
filelock==3.16.1
google-ai-generativelanguage==0.6.10
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai>=0.4.0
googleapis-common-protos==1.65.0
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
identify==2.6.1
idna==3.10
Jinja2==3.1.4
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
nodeenv==1.9.1
oauthlib==3.2.2
pillow==11.0.0
platformdirs==4.3.6
pre_commit==4.0.1
proto-plus==1.25.0
protobuf==5.28.3
pyaes==1.6.1
pyasn1
pyasn1_modules
pydantic
Pygments==2.18.0
pymongo==4.10.1
pyparsing==3.2.0
PySocks==1.7.1
python-dotenv==1.0.1
python-multipart==0.0.17
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==1.3.1
rich==13.9.4
rsa==4.9
shellingham==1.5.4
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.2
tqdm==4.66.6
tweepy==4.14.0
typer==0.12.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
virtualenv==20.27.1
watchfiles==0.24.0
websockets==13.1
groq
streamlit
streamlit_chat
langchain
sentence_transformers
langchain-community
pypdf
boto3